floats instead of doubles

+++ include/ezGfx_mat3.h
@@ -5,7 +5,7 @@
 
 typedef struct {
 	unsigned int rows, cols;
-	double v[3][3];
+	float v[3][3];
 } EZ_Mat3_t;
 
 void EZ_mat3_mul  (EZ_Mat3_t* res, EZ_Mat3_t* A, EZ_Mat3_t* B);
@@ -15,9 +15,9 @@ void EZ_mat3_vdown(EZ_Vec_t* res,  EZ_Mat3_t* A, EZ_Vec_t* v);
 void EZ_mat3_copy (EZ_Mat3_t* res, EZ_Mat3_t* A);
 void EZ_mat3_add  (EZ_Mat3_t* res, EZ_Mat3_t* A, EZ_Mat3_t* B);
 void EZ_mat3_sub  (EZ_Mat3_t* res, EZ_Mat3_t* A, EZ_Mat3_t* B);
-void EZ_mat3_scale(EZ_Mat3_t* res, EZ_Mat3_t* A, double B);
+void EZ_mat3_scale(EZ_Mat3_t* res, EZ_Mat3_t* A, float B);
 
-double EZ_mat3_det(EZ_Mat3_t* A);
+float EZ_mat3_det(EZ_Mat3_t* A);
 void EZ_mat3_inv(EZ_Mat3_t* res, EZ_Mat3_t* A);
