+++ include/ezGfx_mat4.h
@@ -5,7 +5,7 @@
 
 typedef struct {
 	unsigned int rows, cols;
-	double v[4][4];
+	float v[4][4];
 } EZ_Mat4_t;
 
 void EZ_mat4_mul  (EZ_Mat4_t* res, EZ_Mat4_t* A, EZ_Mat4_t* B);
@@ -15,10 +15,10 @@ void EZ_mat4_vdown(EZ_Vec_t*  res, EZ_Mat4_t* A, EZ_Vec_t* v);
 void EZ_mat4_copy (EZ_Mat4_t* res, EZ_Mat4_t* A);
 void EZ_mat4_add  (EZ_Mat4_t* res, EZ_Mat4_t* A, EZ_Mat4_t* B);
 void EZ_mat4_sub  (EZ_Mat4_t* res, EZ_Mat4_t* A, EZ_Mat4_t* B);
-void EZ_mat4_scale(EZ_Mat4_t* res, EZ_Mat4_t* A, double B);
+void EZ_mat4_scale(EZ_Mat4_t* res, EZ_Mat4_t* A, float B);
 
-double EZ_mat4_det(EZ_Mat4_t* A);
-void EZ_mat4_inv(EZ_Mat4_t* res, EZ_Mat4_t* A);
+float EZ_mat4_det(EZ_Mat4_t* A);
+void  EZ_mat4_inv(EZ_Mat4_t* res, EZ_Mat4_t* A);
 
 void EZ_mat4_setId(EZ_Mat4_t* res);
 void EZ_mat4_setProj(EZ_Mat4_t* res, float fov, float aspect_ratio, float z_near, float z_far);
